backend: Add validations for InstallVds (#769196)
https://bugzilla.redhat.com/769196

Adds validations to InstallVdsCommand.canDoAction for provided ISO file
in case of upgrade or re-install of oVirt node:
1. Existence of provided ISO file.
2. Version compatibility - ISO version should have the same major
version as the installed version on the oVirt node.

Since the InstallVdsCommand might fail in its validation when upgrading
rhevh hosts(CDA), there is a need to evaluate the result so updates done
in UpdateVdsCommand could be rolled-back.

Change-Id: Ib5799cbadc9bad36f9f069e8694821a044129c48
Acked-by: FirstName LastName <email>