backend: remove a task from DB only when VDSM clear succeeded (#744823)
https://bugzilla.redhat.com/744823

When a VDSM clear command fail we still remove the task from db. This
results a situation where a failed-clear task is re-inserted to the engine
task manager as Unknown and cannot be cleared.

The solution is to remove the task from DB only when the VDSM clear command
succeded and to leave it clearFailed if not, that way it will attempt to be
cleared from VDSM again by the the schedualed task of the task manager

Change-Id: I8d5c866a9703629fa973dea7037ef457e53cd911
Acked-by: FirstName LastName <email>