ovirt-engine:Report success when restoring snapshot (#743574)
https://bugzilla.redhat.com/743574

When VDSM is restarted, and SPM is re-elected,
All tasks are aborting, and the taskResult returned contains
'cleanSuccess' instead of 'success'.

Since backend didn't get 'success' in the taskResult, we assume that
there was a problem, and we audit log a failure message although delete
image can not be rollback.

The proposed fix is to audit log a success message, even though we get
'cleanSuccess' from the VDSM. (the same behaviour as removeVmCommand
which
use roll forward)

Change-Id: I743ff11462fd67618194ed3f2622517e95a0614e
Acked-by:Omer Frenkel <ofrenkel@redhat.com>