Hard-coded uuid-ossp.sql problem fix.
This patch looks to address BZ #750626

There was a problem with the hard-coding of the location of the required
uuid-ossp.sql file, users without Fedora machines would not have the
ability to specify the location of this file manually. This patch set
adds to the functionality allowing three different options to happen-

1) File is in default location, script executes based on this.

2) User is allowed to specify custom location with '-f' parameter, if
the script finds the file, script executes with this location.

3) File is not in default location and no custom location is specified
with -f. Script fails and alerts user to either install the package
(postgresql-contrib) or specify the location manually.

This is the third revision of the patch. Initial submition was rejected
due to the patch causing execution to stop and wait for user input.
The basis of this original submission was not realising the need for
batch operation, this has since been rectified.

I've changed it now so that it allows another variable to be specified
(-f) to the location of the uuid-ossp.sql file. Otherwise it will look
in the default (Fedora) location of /usr/share/pgsql/contrib/uuid-ossp.sql

Failing that, if it doesn't find anything in the 'f' parameter or in
the default location, it will fail and advise user to make sure file
is either installed, or is specified with -f.

Change-Id: I101338cc8cb685583277049a08f1aad06f37ad97