engine: Adding new TransactionSupportTest with mocking of TransactionManager
The current TransactionSupportTest is problematic in that it does not mock
the external dependencies and requires special setup. Additionally it is
largely a test of the transaction manager implementation, not a test of
the logic in TransactionSupport.

This new set of tests completely mocks the TransactionManager and Transactions
so that the logic in TransactionSupport can be tested independently.

Change-Id: Iac27f96e0642dd8fc60ee3d00f7fcb9df74c4c45
Acked-by: FirstName LastName <email>