engine-core: prevent hard coded timestamps to support other localization settings
Hard coded timestamps cause failure of db creation on other localization settings (Japanese for example)
replace fixed '01/01/2000' date casted to a timestamp with a generated timestamp.
Using to_timestamp function to convert to the locale timestamp from a given format.

(I have preserved the original 01/01/2000 setting since it seems to be complicated to change that in ETL)

Change-Id: I2ab583ac908ea9187f7e48fe29d59cc6dee7a460
Acked-by: FirstName LastName <email>