engine : Restrict size of internal thread pool
The following change was done due to the following reasons:
Under a load circumstances the engine can create an unlimited number of threads which will cause to failure of the system.
In order to do handle a failure of the system was decided to restrict a size of the internal thread pool.
The following change should restrict a size of internal thread pool from unlimited to ConfigValues.DefaultMaxThreadPoolSize.
The minimal size of the thread pool will be ConfigValues.DefaultMinThreadPoolSize, also the max number of the waiting tasks for
thread pool will be limited to ConfigValues.DefaultMaxSizeOfWaitingTasks
The changes were done base on explanation from javadoc: http://download.oracle.com/javase/6/docs/api/java/util/concurrent/ThreadPoolExecutor.html,
the choosen strategy for queuing is "Bounded queues".

In order to do this the following change was done:
1. The changes at InternalThreadExecutor
2. Added configuration options

Change-Id: I182a9227b88f0dba532ce08cd7906209fa12b148