webadmin: events on subtabs dont work (#754085)
https://bugzilla.redhat.com/754085

The Events dont work on any subtab. The problem is, that the main tab
has a timer, which refreshes the main tab, and every time the main tab
is refreshed, the OnEntityChanged method is called. The subtab also
have its timer, which refreshes the events in the subtab. The reaction
of the subtab on the OnEntityChanged is that it stops its timer and
start it again (but without waiting the previous one really
fires an event so the data can really be refreshed).

This fix ensures, that the subtab will only restart its timer, when
not a new instance of an entity arrives but really a new entity
(based on IVdcQueryable).

Change-Id: Ic11a2a1f877dd9c0dd6f5e347f46c64539ece81c