webadmin: Users->General subtab refresh
On the users tab the General subtab. When something changes in the DB,
the main tab is refreshed and shows the correct data, but the General
subtab is not refreshed so shows old data. The consequence is that the
main tab and the general tab shows different data at the same time.

This fix adds a listener to the changes of the data in the main tab
and refreshes the subtab when new data are set.

Change-Id: Ia6ead3b26caf7f5c80dc37830d0f06f730b49295