webadmin: No paging is available through web-admin (#745000)
https://bugzilla.redhat.com/745000

On the web admin it is not possible to page through pages - it's always
 greyed out.

The problem was that the paging is handled in the
UpdatePagingAvailability() method which was called from the
ListModel.setItems() - more specifically this method called the
ItemsChanged() which took care about calling the
UpdatePagingAvailability(). However, in the SearchableListModel the
setItems has been overridden and the ItemsChanged is not anymore called.

The solution was to add to the overridding SearchableListModel.setItems
method also the call of the ItemsChanged(). An alternative solution
would be to call the UpdatePagingAvailability() directly.

Change-Id: Ic933e2035e01ef9bc0bb530f2e87d09adc8c2e2c