backend: Migrate Disks data from Images
Populate the disks table with the disk data from the images table, and
change references so that minimal code changes are needed (at this
point) while the new table is the one being referenced in the views and
STPs.

Later commits should fix the code which addresses the DiskImage entity
to address the Disk entity directly and remove the abstraction which
this commit adds.

Change-Id: Ibea7a6ed33a0d395fdcbe433a900f2a27d00a541