engine-core: setup networks
Edit multiple networks in one command:
e.g. add an eth and bond an existing eth together, or
detach an eth from a bond and move it to another bond.

Also, use connectivityTimeOut to make the VDSM check
the new network topology is functional and the managment
network is visible to the engine.

Change-Id: Ifcc807d01132e8ead95d8104d1a9803ceb591730