webadmin: grid-scroll "jumps" to selected item (#746681)
https://bugzilla.redhat.com/746681

Sometimes the grid-scroll "jumps" to the selected item in the grid,
when the selected item in the grid is not viewable.

The problem was, that when the selection changes, the
SelectionChangeEvent is fired, and as a rection to this event, the
MainContentView updates the whole panel by clearing it and than again
adding the main and detail panels. By this, the selected row losts
it's focus. The previous solution was that the ActionCellTable listened
on browser events and after every, set the focus to the selected row
which caused more problems than just the one described in this
bug - also when the mouse was not in the table no key schemas (such as
key down, shift page up etc) did work.

The solution was to remove the setFocus(true) from the onBrowserEvent2()
and override the onLoad() method of the ActionCellTable and do the
setFocus(true) there.

The second problem solved in this patch is that the
OrderedMultiSelectionModel handled the arrow up and arrow down keys a
bit differently than the HasDataPresenter (which is the presenter of
the ActionCellTable). Say, the first row is selected and the arrow UP
key is pressed. HasDataPresenter considers, that the first rows stay
selected while OrderedMultiSelectionModel considers the last row has to
be selected. The same apply for the last row. This incosistency was
solved by overridding the getKeyboardSelectedRow() method and return
the getLastSelectedRow() from the OrderedMultiSelectionModel.

Change-Id: Ic6c1d274b546e72a37603b8d53268abf4ade18b2