engine-core: Removal of Class File from Compat Package
This patch removes the class File from the backendCompat package.
The patch moves the functionality that is not already there, to FileUtil,
and replaces references to the File class with the FileUtil class.

The patch also removes the relevant test class, and creates the FileUtilTest
class to maintain unit test coverage.

Change-Id: I41709d803eaf7a2220afdf316f5ff8222ffc9d64