engine-core:Add check for locked domains.(BZ751794)
https://bugzilla.redhat.com/751794

while trying to remove storage pool,
The master storage domain is deactivated and then Data Center should be
removed.
Between those two operations, the master domain can be in locked status,
and the storage domain should not be removed.

I added a check if the domain is at lock status, and if so, the
canDoAction should fail.

Acked By: Omer Frenkel  <ofrenkel@redhat.com>

Change-Id: I05c6f193592d06c8073a0f29cbbe3f1d8c015bf6