tools: New tool to upload OVF archives (#782182)
https://bugzilla.redhat.com/782182

The tool provided in this patch makes it easier to upload an
OVF archive to an export domain.  An OVF archive is simply a
zipped archive that can contain an image and must contain
an XML document decribing the image to be uploaded.

-- Tool has been updated to support python 2.6
-- Sample .ovf file has been added to facilitate testing
-- Made changes to deal with conflicting MACs
-- Made changes to handle upload of "Templates"
-- Updated for "sparse" images
-- Tool handles N disk and M snapshot scenarios.

The tool has the following behavior:
1. Before unpacking the archive it will check for requisite space
on the local system.
2. Before uploading the requisite parts in the archive it will
check for space in the target NFS export domain.
3. At this time only NFS as a transport mechanism is supported.
This is slightly different behavior than the iso uploader which
supports both NFS and SSH/SFTP.
4. The tool will allow you to rename the image.
5. The tool will allow you to change the UUID of the image.
6. The tool will allow you to change the disk ID(s) in the image.

Example usage:
1. > python engine-image-uploader.py  -n 127.0.0.1:/virt/exports --template-name=new-name-here upload windows.ovf  --force
2. > python engine-image-uploader.py --conf-file=./imageuploader.conf list
Please provide the REST API password for RHEV-M (CTRL+D to abort):
Export Storage Domain Name     | Datacenter                | Export Domain Status
ExportDomain                   | LegacyDC                  | active
3. > python ovirt-image-uploader.py -e ExportDomain --template-name=new-name-here upload windows.ovf  --force

Change-Id: I3679b44720919aecef457536cc9e19370531c06c