webadmin: add force-remove data-center button (#754417)
https://bugzilla.redhat.com/754417

On the DataCenter tab the force remove button has been added.

The change in the ConfirmationModel is a small bugfix: setIsValid() on
EntityModel fires an event, which is than catched in
UiCommonEditorVisitor. This visitor expects, that the invalidityReasons
are already filled. That's why they has to be set before the setIsValid
is called.
This fix is a part of this push because this bug prevented the
confirmation dialog of the Force Remove to fill the invalidity reason
correctly.

Change-Id: Ic30018714c00dc86511a60e1f781915b880721ab