webadmin: Event main-tab Message column with tooltip (#750292)
https://bugzilla.redhat.com/750292

Since table columns have fixed width and cannot be resized, there is
a chance that the column content will be clipped for some columns.
For example, in Event main-tab table, when switched to 'Advanced View',
the Message column is too narrow to show the entire message. The user,
however, should be able to see the full message.

This patch introduces a new column (TextColumnWithTooltip) which is
similar to GWT TextColumn with following difference:

When the mouse cursor hovers over the table cell, the cell calculates
whether its content fits its area, or if it overflows (is clipped).
In case the cell content overflows, which means the user sees only
a portion of the entire text, a tooltip will be shown to the user when
moving mouse cursor over the table cell.

Table cells whose content fits into their area don't have tooltips
assigned.

TextColumnWithTooltip is used in MainTabEventView, but could be used
in any table-based View that has text columns which could potentially
contain very long strings.

Change-Id: I0ae97cef7926914678cda6ccc3e8d43b625a8afb