webadmin: After logout and login the console icon misses (#767902)
https://bugzilla.redhat.com/767902

The problem was in AbstractActionPanel which in addActionButton method
registered a selection change listener on the model and fired update on
the buttons after catching such an event. However, after logout and
login the model changes, so this listener will never again catch an
event so the action buttons will not be updated.

The fix is to add the CommonModelChangeHandler to re-register the
selection change handler after the common model changes.

Change-Id: Iccf03d98facbed30993e4a9230731bfe6f3b9511