ovirt-engine :  [1/2]Saving upgrade metadata in database & perform only needed upgrades
This patche stores upgrade metadata inside the DB and
run an upgrade script only if it wasn't already intsalled.

The patch adds a first upgrade script that insures that the metadata table used
to store upgrade metadata is installed as first upgarde step.
This patch creates a metadata table for db upgrades
The table stores version number, upgarde script name , checksum of upgrade file,
who installes the script, start and end of script execution and a state of INSTALLED
or FAILED for each script.

Change-Id: I75ffa0fd5d9d4973a32e8c46c19cc74e8cf2d6d3