engine-core: Add Json Serializer and Deserializer MixIn classes
This commit adds mix-in classes for ValueObjectMap references, that
shouldn't be serialized/deserialized. The ValueObjectMap property will
be removed soon, this is a temporary workaround.
Also, it fixes calling the wrong method in the deserializer (some
existing MixIn classes were added to the Serialization configuration
instead of the Deserialization one).

Change-Id: I1a62e11ffe0faed814a324ef5f4b3e2757bba10d