core: [2/2]Do not enable upgrade script to run if there are upgrade files with duplicate version
Do not allow duplicate versions in upgrade scripts.
Version is considered as the 8 digit prefix of the file name
For example if file is 03_01_0999_descriptive_name.sql than the version
associated with that file is 03010999

In the case that files with same version exists, upgrade aborts
reporting the duplicate version found.

Change-Id: Iaf70d62fb38f1f920dbeb2282819fdd5ed5026fa