core: Adding scripts to check db schema
The following patch starts adding the mechanism of
checking correctness of db schema.
Upon running the upgrade script, the schema_version table
holds md5 values per each excuted script.
An md5 calculation on all the md5 values is performed and the result
is stored in a file.
Prior to jboss execution, it will be determined of the value stored
in the file matches a calculation that will be run against the db.
This patch introduces the db schema md5 calulcation ahd the check.

Change-Id: Ib0668b28a41903c22895bceb2bf85b2f5c3c2b8e
Acked-By: