webadmin: Enter key closes the dialog, instead of excuting 'GO'(#770417)
https://bugzilla.redhat.com/770417

Click on "Configure" button -> choose the "System Permissions"
tab -> click on "Add" -> write a name to search and press on Enter, the
dialog will be closed although the expected behaviour is that the "GO"
button will be on focus (the"GO" button will be triggered by
pressing Enter).

The problem was that the enter is handled by default as
 model.getDefaultCommand().Execute() which in this case is the OK
button of the popup and not the GO button.

This fix decorates the handler performing this default behaviour. The
decorator executes the GO if the enter has been pressed and the focus
is on the search text field, othervise it forwards the event to the
default handler.

Change-Id: I98c788313032b138cded2bb918637a44966398a7