webadmin: System trees object names appear one line below icon (#769875)
https://bugzilla.redhat.com/769875

Move the System tree's right-hand pane window to the left, the names of
objects in the System tree which "touch" in this border will appear one
line below their icons.

This fix sets white-space: nowrap; for ScrollPanel and CellTree into
HorizontalPanel which forces the scroll bar to appear when the text
overflows. Unfortunately I did not find a way to solve this without
using tables.

Change-Id: Ia9c9703342ce26f3cc83d74c5452f9f58695b800