core: Task Manager Add Tables
The patch adds 3 tables for the Task Manager:
Job table - stores the job entity, describes an action in the system.
Job_subject_entity_map - stores the relation between a job to its
subject entities.
Step - stores the step entity, a part of the job.

More specific data about the entities/tables could be found in
http://ovirt.org/wiki/Features/TaskManagerDetailed

Change-Id: Ib662f58d8a739b1fa44f4c4ec25f053923178fc0