core:auto generate .schema file in case of upgrade changes
Currently it is hard to see the current database schema in one place since we have a
base database and a bunch of running upgrade scripts that changes it.
This patch uses pg_dump utility to generate the schema after an upgrade change occurs
and stores it in a .schema file.
The .schema filke is local and is added to .gitignore
Now we can look in one file for the current schema, this is very helpful when doing any change
that should consider references and depndencies.

Change-Id: Ic4a3de856e30758419634b18364c286bb7798d11