webadmin: installed (failed) hosts will not appear in the main grid (#772674)
https://bugzilla.redhat.com/show_bug.cgi?id=772674

When there are Hosts with InstallFailed status, there is a NPE thrown.

The Exception is thrown when there is a Host in InstallFailed status.
This is caused by the fact that the vds.getnet_config_dirty() property is null.
This property is a Boolean rather than a primitive boolean, and it seems that
GWT does not handle the Unboxing properly when the Boxed value is null.

Change-Id: I721ca908fa503fcebe42e5a8325d68f3e3a362ff