engine-core: removing restriction on adding disks to different storage domains
Removed check in canDoAction in AddDiskToVmCommand so that a disk can be
created on a different storage domain than its VM. Added code to
AddVmTemplateCommand to use the storage domain id from each image. Changed
ImagesHandler to use the disk image storage domain id if it is available.
Also removed uneeded method and BllMessage

Change-Id: I0c36b87991e2ebb78ea46ca3beb44818e40e1806