webadmin: Dialog disappears: it is impossible to re-open it (#772609)
https://bugzilla.redhat.com/772609

When the popup is opened and the hash fragment of the URL changes, the
popup disappeares and makes it impossible to open any popup
in the specific tab again.

The problem was, that on the AbstractPopupView the
setAutoHideOnNavigationEventEnabled(true) was called.
This caused, that when the hash fragment changed, the popup's hide()
method was called to hide the popup. As in WebAdmin, the logic of
hiding or showing the popups is driven by the uicommon model objects,
this caused incosistency between what was shown and what the model
expected that is shown. In this case, the model expected, that the
Dialog is still visible and it was not.

As the only component responsible for such logic is uicommon, the
setAutoHideOnNavigationEventEnabled has been removed.

Change-Id: I62fd06130c6a689997ecfa7d5e14aa3b19683cdb