webadmin: IdHandler support for extending IDs at runtime
Added new method to ElementIdHandler interface:

void setIdExtension(String extension);

Calling this method prior to generateAndSetIds(owner) will cause
generated DOM element IDs to be "extended" with the given value
(the extension happens at runtime).

This can be helpful when there are multiple instances of the same
owner type displayed on the page. In case of WebAdmin, this can
be helpful to distinguish between IDs of two same dialog Views
displayed at once, avoiding potential ID clashes.

Runtime ID extension is based on the following assumption:
Extending already unique ID will result in another unique ID.

It is up to the user to extend IDs for owner types which will
have multiple instances shown on the page during runtime (IdHandler
cannot determine how classes will be used at runtime, but it can
provide an API to help with this scenario).

Change-Id: I92d2bad4680e31829a84de977bcf44ecfddf8051