webadmin: Clicking on Enable Power Management does nothing (#771893)
https://bugzilla.redhat.com/771893

When a host has no Power Management configured, clicking on the
Enable Power Management action item does not do anything. It should
open the Edit Host dialog on the Power Management tab.
This is true for all host statuses.

There were two problems:
- After logout-login still the old model instance was used in
SubTabHostGeneralPresenter. This caused that when clicking on the action
button nothing happened. Solved by listening on onCommonModelChange()
and re-initializing the alerts.

- The check, if the power management tab should be opened or not was
done before the data from the server arrived
(in HostPopupPresenterWidget.init()). This caused that always the
General tab has been shown. The fix was to move this check
into a listener, which waits until the relevant data from the server
arrives and after than activates the relevant tab.

Change-Id: I61c04f2dafb53f3558a4134e31d61b3d80c25032