core: Task Manager Add Job Support to Multiple Action
The patch adds support for the multiple action runners. The default
behaviour for external call for multiple action runner is to create a
monitored Job entity. For internal multiple action invocation, the
default is not to produce a Job for it, where the option to override
this default is provided by calling new API which gets as parameter the
execution context by which the command should be executed.

Change-Id: I50bc6d0cb43468dd3803101eeaf77256080fee06