core: Task Manager Add StepId property to async_tasks entity
The patch adds the step ID property to the async_tasks entity and to its
persistent objects (table, stored procedure and DAO) and modifies the
calls for the c'tor with extended one.

The step ID is a nullable field which store the ID of the referenced
step that describes the task status.

Change-Id: If83edba44f6aad23eab8dddd816b00fbcd1ff00c