core: Exit in case of an error
Adding a parameter to execute-file to determine if we want to exit or just return
the error code.
when an error occurs, script execution is halted except on the upgrade process in which
we will have also to update the schema_version table and in that case we will only
return the code.

Change-Id: I14000f3ce84f0f7acbae66240ccce6d1b4834934