core: Simplify RemoveSnapshot command to get id
The RemoveSnapshot command will get the ID of the snapshot to remove,
and will take care of figuring out who to merge to by itself.
This is agreeed with REST/UI behaviour and also in accordance to planned
changes on VDSM layer which is planned to do the merge of images
internally.

Change-Id: I7b79071550b4b53774ea9482a0ca2eda2f65b1fd