core: Added ValidationResult & handling
Added ValidationResult class to help pass result of validation that is
done for canDoAction methods.
This way, the canDoAction method can simply call a validator using:
  validate(Validator.doSomeValidation(...));

This will help write simple and concise validators that will not need to
be aware of how command base handles canDoAction internally, while
keeping the calls to them simple.

Change-Id: I755072c4217a2ea204ab9f3f79b81c34968cf2f5