core: Task Manager VDSM tasks monitoring - cont.
The patch passes the execution context to internal commands which
invokes VDSM tasks, so that the tasks could be monitored as steps,
related to the main Job which was created by the action.

There are cases in which the internal command calls an internal command
which creates the VDSM tasks either directly or by invoking an internal
command, therefore the execution context object is
being provided for it as well.

The patch uses the CommandContext as the context object for the internal
backend API calls.

Change-Id: Ib90ccce5bd2fc318235d3a6ce096afca5ac99ff0