ovirt-engine: Add quota usage functions
In the next patch I added quota functions to calculate the usage of the
quota in real time.
The functions that are being added are:

CalculateVdsGroupUsage - Summerise the VCPU and memory usage for all the
VMs
in the quota which are not down or suspended.

CalculateStorageUsage - Summerise the storage usage for all the disks in
the quota, for active disks, we summerise the full size,
for snapshots and other disks summerise only the actual size.

Acked By: Eli Mesika  <emesika@redhat.com>
Change-Id: Ifa6caf21f3170b8e69a7a154a483d605b71e0327