core: Task Manager Support monitoring of VDSM tasks
The patch adds support for monitoring VDSM tasks as steps of a job.
Each task is represented as a step. Upon task completion the step will
be updated to reflect its status.

A default implementation provided for internal commands which
invokes commands which creates VDSM tasks directly or by invoking
internal commands.

The tasks are being marked as completed by the SPMAsyncTask which
handles ended tasks.

Upon an invocation of the command EndAction (by EntityAsyncTask) the Job
entity will be updated for completion with the relevent status.

Change-Id: I3821ab99fe3d9a7ff5e977456df0aa50fe1cfc9e