core: Task Manager Add ExecutionContext
The patch introduces the ExecutionContext class which acts as the
monitoring execution plan, the guides by which the commands in the
system should be monitored.

A commands can be run as a Job or as a Step (specifically internal
commands). The monitoring level are defined by the ExecutionContext which
is initialized with default implementation guidelines:
1. External commands are monitored
2. Internal commands aren't monitored
3. Internal commands with VDSM tasks will present a VSDM task as a step
(implementation in further patchs).
4. Job has 2 steps by default: VALIDATING and EXECUTING.

Change-Id: I680667682cc80bcc383e22d8da0d4d340184c020