core: Task Manager Add JobRepository
The patch adds the JobRepository interface and its implementation,
responsible for manipulating the Job entities using the job DAOs.

It encapsulates task manager related operation as a single API and
define the atomicity of the operations. DML operation are executed in a
new transaction, detached from the scope in which the command is
running.

Change-Id: Ie35176f042bfd8d21100a76e3ac7a4f4dcb9c332