core: Add ExecutionMessageDirector for job messages
The patch adds the ExecutionMessageDirector, a class which is
responsible for reading the job/step messages from a resource bundle by
a key identifies the message.
.
The messages keys are categorized as Job or Step level:
1. Job level: job.VdcActionType=message
   The job key combines the "job." prefix with the action type which the
   job describes.
2. Step level: step.StepEnum=message
   The step key combines the "step." prefix with the step name as appears
   in the StepEnum.

The resource bundle for the message is provided as well, were few
actions and steps are described. The complete entries will be provided
in a separate patch.

Job and Step without an message will be presented as the name of enum
which is associated with them.

The message format is a text message with place holder to be replaced
with a value, such as ${AnyParam}. The provided keys should be lowered
case.

Change-Id: I567d547608e6825336fad91e14d11c5f73fa8024