core: Task Manager Add CommandBase.getJobMessageProperties()
The patch adds a new method to CommandBase for resolving the properties
to be set for the Job description of the command.

The default implementation try to resolve the
PermissitionSubjectEntities for the specific entities which are used by
the command and get their value from the DB. If failed to resolve the
value, the entity ID will be set as the property value.

Change-Id: Iba069f715202890276c48c903260fb8c519249c1