core: Replace InstallVds invocation method in AddVds
The patch replaces the Backend.runInternalMultipleAction() with
Backend.runInternalAction() invoked in a separate thread from the one
uses by the AddVdsCommand.

The Backend.runInternalMultipleAction used before to achieve
asynchronous invocation, however it abuses the multiplicity purpose of
that API.

The change will provide the ability to pass an execution context to the
InstallVdsCommand in more appropriate way.

Change-Id: Idbb788c4396c8fdbe183e4508770137496b0d454