engine: Create default unlimited quota for DC
In the next patch, I added a creation of  new default unlimited quota
every time a new DC is created and when ever a DC will become disabled.

We create new quota when ever DC become disabled since we can't be sure
if the user didn't changed the default quota.

Change-Id: I10db583cca37f8edd6894aba5053f862dc96af65