webadmin: hosts tab- pm config window is opened when editing host
https://bugzilla.redhat.com/show_bug.cgi?id=784522

Hosts tab: when editing host, general tab should be focused and not pm tab.
When opening the window from the "Enable Power Management" link in the
"General" sub-tab, the Edit Host Window should be opened with the
"Power Management" section in focus and with the "Enable Power Management"
check-box checked and disabled.

Change-Id: I26a0df34b84b5c5ce8c9afc11cff5a82b55e6efd