core: Introduce notion of internal queries
currently, this is only a property (which can be set and queried)
applied to a query object.
Backend will set it when calling runInternalQuery.
Note that this is currently a "stuby" implementation for the query
permissions mechanism, and does not effect the execution of the query
(yet).

(Already approved by mkublin - resent due to merging issues, no code
changes)

Change-Id: I11e9977b58cee2d919c44d6940daeee2a3ff7da3