engine: Add default quota indication field
In the next patch, I added a column to the quota table which
indicates if the quota is default in the Data Center.

The default quota is an unlimited quota which resources are consume from, when the DC
configured as quota disable.
Default quota, should be created when ever DC will be created and when ever DC
will become disabled (This change can be seen in an advenced patch as
part of this patchset).

Change-Id: I2f99169114a41ffcb7e06609d43623711cfced99