core: Refactor of VdsDynamic.getSupportedVersion
The patch refactors the VdsDynamic:
It removes the usage of the LogComapt from within the class to loose the
dependency on the Compat project (and specifically on the LogCompat).

By that, the possible thrown exception as a result of the version
parsing will be handled by the caller for the methods:
HandleVdsVersionCommand.

Since VdsDynamic.getSupportedENGINESVersionsSet and
VdsDynamic.getSupportedClusterVersionsSet shares the exact logic, a
common message was extracted.

Change-Id: Ic8d0d9ac09cdbca563515774382711fad4913452