core: Adding version dependant config values for version 3.1
vdc_options includes settings marked with version='general' that indicates that
this setting is not version dependant and with other settings marked with version='x.y'
where x is the realesae number and y is the major number of the version.
Those entries indicates version dependant values and are repeated with the correct
setting for each added version.

V2: Fixing EmulatedMachine to Fedora

Change-Id: I6bbc61210d175f178d2d1974b1f09944bff8c070