core: Prestarted Vm (2)
This patch adds the basic new flow to the AttachUserToVmFromPoolAndRunCommand.
The command first tries to find prestarted Vms, and if succeeds in doing so,
only runs AddPermissionCommand. If a prestarted Vm is not found, we need to
also run the RunVmCommand.

It also moves Vms in pools from working with the image_vm_pool_map table
to using the stateless table for images. In the new flow, when adding
a Vm to Vm Pool we do not need to add its images to any table (but rather
only when running the Vm).

The patch also changes the behaviour of RemoveVmFromPoolCommand.

This patch adds the logic to find a prestarted Vm, when
coming to assign a Vm from a VmPool to a user.

It also cleans image_vm_pool_map from BaseDAOTestCase.

Change-Id: Ie3c0b0d69c0da82d3ba2dcbcf045cd92a6db7bef