engine: Add quota manager class
In the next patch I introduced quota manager.
Quota manager responsibility is to encapsulate the validating process
for quota for the commands to use it simply.
The quota manager synchronizes requests for resources with a memory delta table, which should
reflect the delta changes, before persisting them into the DB.

Every operation that consume resources cluster wize or storage wize
should be go through the manager class mechanism.

Change-Id: I491273277209cdc34072786a298a59d93541afd4