engine: Adding fix to quota stored procedure
In the next patch I changed the stored procedure to retrieve quota
limitation id even if there is only global limitation and not specific
one.

This fix is needed for the Quota Manager in the next patch, since it is
using the stored procedure to save the delta status for each quota.

Change-Id: I95d3683f820bb23c51d3ead6caa6783cb5f88340