webadmin: Simplifyed generics on editors
When the editor delegates are generated, the GWT assignes them a name
like this:
editorName_genericArg1_genericArg2_..._UiCommonModelEditorDelegate.java
As the genericArg is a fully qualified name of the class which is the
 generic argument, the length of the filename can grow to be quite big.

As the filesystems has a limit to the length of the files names
(e.g. ext4 can have max 256B long filenames only) it can happen that
the size of the file can exceed the allowed length for the specific
file system.

This was a particular problem with WidgetWithLabelEditor when when I
needed to generalize the EntityModelCheckBoxEditor to
BaseEntityModelCheckBoxEditor, the resulting file name looked like this:

WidgetWithLabelEditor_org_ovirt_engine_ui_uicommonweb_models_vms_
ConsoleModel_org_ovirt_engine_ui_common_widget_editor_
TakesValueWithChangeHandlersEditor_org_ovirt_engine_ui_common_widget_
editor_BaseEntityModelCheckboxEditor_UiCommonModelEditorDelegate.java

which is too long to be stored on ext4 (or ext3, NTFS, FAT...)

The solution was to simplify the WidgetWithLabelEditor to have less
generic arguments so the resulting filename will be shorter.

This patch contains only the generic arguments simplification, not
the generalized BaseEntityModelCheckBoxEditor.

Change-Id: I10354757de370e0b7b036c52150758bc64bfa2f2