core: Prestarted Vm (3)
This patch adds an upgrade script to migrate the data in
image_vm_pool_map to the stateless_vm_image_map table.

It also cleans up the image_vm_pool_map in general.

In addition, a minor change was made to the
AttachUserToVmFromPoolAndRunCommand to go with the new
Task Management.

Change-Id: I53059691c0d9c3e9bdc6cda799f7fd1863981a79