userportal,webadmin: Replace using LogCompat
The patch proposes alternative of using the LogCompat by GWT components.
The LogCompat was removed from the Compat project, and any usage of it
was removed from the Common project.

The LogCompat is deleted from the overridden files, and replaced by a
other file which overrides the org.apache.commons-logging.

Adjustment were made for searchbackend which relied on previous
interfaces of LogCompat.

Change-Id: I22f54c630f6df2a46e2e1261344037d967607a5b