webadmin: moved the webadmin specific definition to webadmin
The UiMenuBarButtonDefinition contained this configurations of the
triangle_down image:
@Source("org/ovirt/engine/ui/webadmin/images/triangle_down.gif")
@ImageOptions(width = 7, height = 5)

As this two options are specific to webadmin, they should be defined on
the webadmin layer. The abstract layer should not define the project
specific resources or create (GWT.create) project specific resources.

The practical consequence in this case was a fail of the build, because
the org/ovirt/engine/ui/webadmin/images/triangle_down.gif was not
found in the gwt-common project

Change-Id: If19bc6669e6f42d94c97489a912a8732ea1bb78b