core: Task Manager Add Jobs description resource
The patch adds the jobs and steps descriptions as will appear to the
user in the Tasks sub-view.

The prefix "job." represents actions which the user has triggered,
correlated to the actions from VdcActionType with an action group.

The prefix "step." represents steps of the job, there are common steps
as VALIDATING, EXECUTING and FINALIZING and also customized steps to
describe specific part of the job.

The bottom section of the properties file describes non-monitored
commands, hence the update for the VdcActionType enum.

The Step section will be update to shed more light about the entities
being processed by the VSDM tasks.

Change-Id: I0244fad5afb56b291e0e4843b03342a579ea7007