core: Prestarted Vm (5)
1. This patch adds a periodic job that maintains the desired
number of Vms in a Vm Pool. It new config values
so the intervals of the job, and the size of the batch of
Vms that are prestarted in every run, can be set by the admin.

2. It also adds a validation for setting the prestarted_vms
field, to make sure the number being set does not exceed
the size of the Vm Pool.

3. The patch changes the flow in VmPoolHandler which
deals with stopping a Vm.

Change-Id: I5b5394c170660431272f74d7015983ddfffab39a