engine: Add quota validation for AddVmCommand
In the next patch, I added support when validating quota at creation of
VM command.
The validation assembled from several phasas, and count on a delta
memory map mechanism, described in the quota manager.

The validation start with the method validateQuota, which then
update the delta map with the size requested to consume.
After the operation finished, we remove the delta that was saved
before from the delta map.

Also added helper method addQuotaToConsumeMap to the QuotaHelper class.

Change-Id: Ia2af7b2a1376db99593c42127b172925d8e88a98